+++ src/hooks/AuthUserService.ts
@@ -1,5 +1,6 @@
 import prismaClient from "../prisma";
 import { compare } from "bcryptjs";
+import { sign } from "jsonwebtoken";
 
 interface AuthRequest {
     email: string;
@@ -32,7 +33,25 @@ class AuthUserService {
             throw new Error("Usuário/senha incorretos");
         }
 
-        return { ok: true };
+
+        const token = sign(
+            {
+                name:user.name,
+                email:user.email
+            },
+            process.env.JWT_secret,
+            {
+                subject: user.id,
+                expiresIn:'30d'
+            }
+        )
+
+        return {
+            id: user.id,
+            name: user.name,
+            email: user.email,
+            token: token
+         };
     }
 }