+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created controller and user service
 - Fixed errors
 - Created validation of users
+- Created auth service and controller
+- added JWT
+- added dotEnv
+- Created token
 
 ### Fixed